extract box layout into internal layout package
Thus it can be extended for menu usage without exporting it.
